+++ datafusion/physical-plan/src/filter.rs
@@ -25,9 +25,9 @@ use super::{
     RecordBatchStream, SendableRecordBatchStream, Statistics,
 };
 use crate::common::can_project;
-use crate::projection::ProjectionExec;
-use crate::projection_utils::{
+use crate::projection::{
     make_with_child, try_embed_projection, update_expr, EmbeddedProjection,
+    ProjectionExec,
 };
 use crate::{
     metrics::{BaselineMetrics, ExecutionPlanMetricsSet, MetricsSet},
@@ -421,17 +421,13 @@ impl ExecutionPlan for FilterExec {
             return Ok(None);
         };
 
-        match FilterExec::try_new(
-            new_predicate,
-            make_with_child(projection, self.input())?,
-        )
-        .and_then(|e| {
-            let selectivity = self.default_selectivity();
-            e.with_default_selectivity(selectivity)
-        }) {
-            Ok(filter) => Ok(Some(Arc::new(filter) as _)),
-            Err(_) => try_embed_projection(projection, self),
-        }
+        FilterExec::try_new(new_predicate, make_with_child(projection, self.input())?)
+            .and_then(|e| {
+                let selectivity = self.default_selectivity();
+                e.with_default_selectivity(selectivity)
+            })
+            .map(|e| Some(Arc::new(e) as _))?
+            .map_or_else(|| try_embed_projection(projection, self), |e| Ok(Some(e)))
     }
 }
