+++ datafusion/physical-plan/src/joins/symmetric_hash_join.rs
@@ -47,9 +47,9 @@ use crate::joins::utils::{
     BatchTransformer, ColumnIndex, JoinFilter, JoinHashMapType, JoinOn, JoinOnRef,
     NoopBatchTransformer, StatefulStreamResult,
 };
-use crate::projection_utils::{
+use crate::projection::{
     join_allows_pushdown, join_table_borders, new_join_children,
-    physical_to_column_exprs, update_join_filter, update_join_on,
+    physical_to_column_exprs, update_join_filter, update_join_on, ProjectionExec,
 };
 use crate::{
     joins::StreamJoinPartitionMode,
@@ -566,7 +566,7 @@ impl ExecutionPlan for SymmetricHashJoinExec {
     /// Otherwise, it returns None.
     fn try_swapping_with_projection(
         &self,
-        projection: &crate::projection::ProjectionExec,
+        projection: &ProjectionExec,
     ) -> Result<Option<Arc<dyn ExecutionPlan>>> {
         // Convert projected PhysicalExpr's to columns. If not possible, we cannot proceed.
         let Some(projection_as_columns) = physical_to_column_exprs(projection.expr())
