+++ datafusion/physical-plan/src/lib.rs
@@ -67,7 +67,6 @@ pub mod memory;
 pub mod metrics;
 pub mod placeholder_row;
 pub mod projection;
-pub mod projection_utils;
 pub mod recursive_query;
 pub mod repartition;
 pub mod sorts;
