+++ datafusion/physical-plan/src/streaming.rs
@@ -24,9 +24,8 @@ use std::sync::Arc;
 use super::{DisplayAs, DisplayFormatType, PlanProperties};
 use crate::display::{display_orderings, ProjectSchemaDisplay};
 use crate::execution_plan::{Boundedness, EmissionType};
-use crate::projection::ProjectionExec;
-use crate::projection_utils::{
-    all_alias_free_columns, new_projections_for_columns, update_expr,
+use crate::projection::{
+    all_alias_free_columns, new_projections_for_columns, update_expr, ProjectionExec,
 };
 use crate::stream::RecordBatchStreamAdapter;
 use crate::{ExecutionPlan, Partitioning, SendableRecordBatchStream};
